Create Count the paths
Given a Directed acyclic graph(DAG) with V nodes labeled from 0 to V-1. Given edges, start, and destination, count the number of ways to reach from start to destination. There is a directed Edge from vertex edges[i][0] to the vertex edges[i][1]. Your task is to find the number of ways to reach from start to destination.